don't assume input vector to be sorted in prefix_*
